Avoid overriding cache-and-network fetchPolicy when refetching.
Should help with the problem reported by @supercranky in this comment:
https://github.com/apollographql/apollo-client/issues/4636#issuecomment-483218323

The apparent reasoning for preserving the no-cache and network-only fetch
policies is that refetching requires a "network" fetch policy, but I
believe cache-and-network should also count as a "network" fetch policy,
since it always attempts to fetch from the network in addition to fetching
from the cache. Even if that network request fails, it may be useful
(e.g. for offline usage) to return data from the cache.